Implement PR feedback, fix typo, add conversion import
